[CPDNPQ-3212] Handle missing mandatory institution answer
When users do not follow the wizard path, we can end up with a `nil` institution for a journey where it should never be `nil`. This can happen if e.g. the user has multiple journeys open in multiple windows with different work settings selected, where the session is written in such a way that they skip required institution picker steps.

We shouldn't ignore these nils with `try` or safe navigation operators, because this will lead to false negative funding outcomes. Instead we need to detect when an institution should be present but isn't, and make sure we get that information.