[CPDNPQ-3212] Handle missing mandatory institution answer (#2874)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3212

When users do not follow the wizard path, we can end up with a `nil`
institution for a journey where it should never be `nil`. This can
happen if e.g. the user has multiple journeys open in multiple windows
with different work settings selected, where the session is written in
such a way that they skip required institution picker steps.

Currently this results in exceptions as linked in the Jira ticket, i.e.
the user gets a dead-end generic error screen.

We shouldn't ignore these nils with `try` or safe navigation operators,
because `nil` is not in fact a legal value in the scenario, and as such
this will lead to false negative funding outcomes. Instead we need to
detect when an institution should be present but isn't, and make sure we
get that information.

### Changes proposed in this pull request

- Make funding eligibility scenarios where an institution is mandatory
explicitly call for one
- Raise an exception if that mandatory institution is `nil`
- In the registration wizard, rescue from that exception redirecting the
user back to the relevant question so that they can continue