[CPDNPQ-2604] Standardise and test external links (#2865)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-2604

- Move external link URLs out of views into `config/external_links.yml`
- Encapsulate these links with the `ExternalLink` class
- Add an `external_link_to` helper to use references for external links
rather than hard-coding URLs
- Add a rake task `external_links:verify` to check that all the links
result in a 2XX HTTP status, or raise an exception if not
- Add a GitHub workflow to run the above rake task

If verification fails, an exception will be raised. In CI, this
exception will be reported to Sentry and alerted in Slack (or Teams
imminently).

The ticket specifies running this weekly, but I went for daily since
this is a very cheap operation and couldn't see any downside in checking
more frequently.

I don't think this should be part of the checks for PRs / merges, since
external services are not under our control and shouldn't block our
deploys. However, I've got action configured to run on Pull Requests at
the moment, since this is required to be able to test it. As detailed in
the Jira ticket, this can be tested by adding a commit that adds a bad
link to `config/external_links.yml` — after that, we'll remove the test
commit and also remove the Pull Request action, so that the action is
run on a daily schedule or by manual workflow trigger only.