[CPDNPQ-2997] Add Application#reason_for_rejection (#2894)
(Supercedes #2881 — new PR due to editing the migration after deploying
review app)

### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-2997

There are several reasons that an application can end up in the rejected
state, which can make it hard for Lead Providers and administrators to
understand. Recording and exposing this information should make the
service more efficient.

n.b. after chatting with Slawosz and Alkesh, it seems that we can't add
swagger documentation that depends on feature flags, so this is a
conscious omission from the Jira acceptance criteria.