[CPDNPQ-2998] Lead Provider to have visibility of declarations if applicants have transferred (#2896)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-2998

- When applicants transfer Lead Providers, the new Lead Provider should
be able to see existing declarations made by the previous Lead Provider
- The previous Lead Provider should continue to only see their own
declarations
- No change to write permissions

Changing the query seems the most expedient route for this, since it's
used in the declaration endpoints, and doesn't affect any write
permissions. With this change, we grant visibility to the current
`Application` owner, as well as the `Declaration` owner.