Record job as the actor in PaperTrail versions (#2935)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3209

If data is changed in a background job, then we do not record the actor
in the whodunnit field. This is useful information which we'd like to
retain.

### Changes proposed in this pull request

Set the whodunnit for all background jobs, with the name of the Job
making the changes.