[CPDNPQ-3058] admin upload eligibility lists only (#2948)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3058

Splitting this ticket into two:

1. allow eligibility lists to be uploaded into the database, but do not
use the new lists - old ones will be used.
2. switch to using the new database eligibility lists, and the code
referring to the old lists will be removed.

This PR is the first step only.

### Changes proposed in this pull request

- page with tabs to allow new eligibility list files to be uploaded
- one new table (eligibility_list_entries) to store all the entries,
with a type using Rails STI

a couple of things to note:
- the disadvantaged EY list has two mandatory columns: `Disadvantaged EY
School URN`, `Ofsted URN` - we use the ofsted URN if it's available,
otherwise we use the other one.
- currently, only the disadvantaged EY CSV is in ISO-8859-1 format -
this PR copes with ISO-8859-1 and UTF-8 encoding.