[CPDNPQ-3288] cope with course being nil during registration journey (#2974)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3288

Navigating directly to certain pages causes an error, because they need
a course chosen in a previous step.

### Changes proposed in this pull request

For steps that need `course` to be set in the form, redirect to
course-start-date (the first question on the journey after logging in).
If the user is not logged in, they get redirected to the start page
anyway.