[CPDNPQ-3192] Fix display of timestamps in admin (#2844)
### Context

Ticket:
[CPDNPQ-3192](https://dfedigital.atlassian.net/browse/CPDNPQ-3192)

We are currently showing the timestamps an hour out in the admin
console. This is because we are configured to show timestamps in UTC,
not BST

### Changes proposed in this pull request

1. Swap to using a UK local timezone which should then display correctly
in both summer and winter - this is only done for web facing requests to
avoid impacting data representation on the API

Note: we also have a lot of specs which assume they are running in UTC,
which is another reason for leaving the Rails.configuration.time_zone as
is.

I did a spike with forcing the APIs to use UTC but the changes were
proving more involved than I'd like and this solves the current issue of
an application from 5 minutes ago showing as being from 1 hour and 5
minutes ago

### Failed feature specs screenshots

If any of the feature specs would fail, there will be page on the wiki
with all
failures:

https://github.com/DFE-Digital/npq-registration/wiki/


[CPDNPQ-3192]:
https://dfedigital.atlassian.net/browse/CPDNPQ-3192?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ