[CPDNPQ-3316] improve contracts download (#3000)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3316

Currently, the contract CSV returns empty columns if the
`monthly_service_fee` is null,
which makes them unusable as-is - they have to be manually altered to be
used.

Also, if there are multiple statements for a given lead provider and
course, we currently return multiple rows representing all the possible
values.
This is not useful, as there should only be one row per lead provider
and course.

### Changes proposed in this pull request

- return `0.0` for null `monthly_service_fee`s
- for each lead provider and course, return the contracts values for the
latest statement that had output_fee=true