[CPDNPQ-3273] Remove legacy API code (#3025)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3273

We are removing redundant legacy API code that is no longer required in
the codebase.

### Changes proposed in this pull request

- All code referencing the config.x.disable_legacy_api key is removed
- All code referencing Rails.configuration.x.disable_legacy_api is
removed
- The config.x.disable_legacy_api key is removed from
config/application.rb

Includes the removal of any test specs for the routes that were removed
along with controller actions for the routes removed.

### Failed feature specs screenshots

If any of the feature specs would fail, there will be page on the wiki
with all
failures:

https://github.com/DFE-Digital/npq-registration/wiki/