accessibility statement changes (#3062)
### Context


https://dfedigital.atlassian.net/jira/software/projects/NPQ/boards/137?jql=assignee%20%3D%2061d42643ce3652006ab6cd2a&selectedIssue=NPQ-3304

We are updating the statement because our levels of accessibility have
changed.

### Changes proposed in this pull request

Remove descriptions of accessibility targets that we have now met and
generally improve the readability of the statement.

### Failed feature specs screenshots

If any of the feature specs would fail, there will be page on the wiki
with all
failures:

https://github.com/DFE-Digital/npq-registration/wiki/

<img width="2256" height="4995" alt="image"
src="https://github.com/user-attachments/assets/d36647a2-69d0-470d-88fc-3ae8ba7eee29"
/>