[CPDNPQ-3323] - sortable table for in review section (#3021)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3323

We want to be able to in admin area, sort the application that are in
review by newest or oldest.

### Changes proposed in this pull request

Created a new sort filter that is a dropdown select box with a butto to
sort applications. All while maintaining the filter options provided to
the user.

https://github.com/DFE-Digital/npq-registration/wiki/