[CPDNPQ-3344] - Update default view on finance/statements page (#3037)
### Context

Ticket: https://dfedigital.atlassian.net/browse/CPDNPQ-3344

The number of statements being shown by default was for all statements
which we felt was too large. This is to reduce that number down on the
first page load to only show statements with a payment run.

### Changes proposed in this pull request

The filter option for payment run now defaults to "Yes" on first page
load to reduce number of records being returned.

https://github.com/DFE-Digital/npq-registration/wiki/