[CPDNPQ-3267] - record rake tasks for manipulating data (#3074)
### Context

Ticket: https://dfedigital.atlassian.net/browse/NPQ-3267

### Changes proposed in this pull request

Adds the ability to record changes made to our database models as part
of a rake task which previously was not the case.

Now when we run one off tasks to manipulate date, we set the whodunnit
field on versions table to the name of the rake task.